requirements.txt
azure-storage-blob
azure-data-tables
azure-identity
configparser
pyyaml
psycopg2